refactor: prepare for middlewares
