+++ transport/transport.go
@@ -1,10 +1,9 @@
 package transport
 
 import (
-	"encoding/base64"
+	"bytes"
+	"io"
 	"io/ioutil"
-	"net/http"
-	"strings"
 
 	"github.com/mitch000001/go-hbci/segment"
 )
@@ -13,42 +12,13 @@ type Transport interface {
 	Do(*Request) (*Response, error)
 }
 
-func NewHttpsTransport() *HttpsTransport {
-	return &HttpsTransport{
-		httpClient: http.DefaultClient,
-	}
+type TransportFunc func(*Request) (*Response, error)
+
+func (fn TransportFunc) Do(req *Request) (*Response, error) {
+	return fn(req)
 }
 
-type HttpsTransport struct {
-	httpClient *http.Client
-}
-
-func (h *HttpsTransport) Do(request *Request) (*Response, error) {
-	encodedMessage := base64.StdEncoding.EncodeToString(request.MarshaledMessage)
-	httpResponse, err := h.httpClient.Post(request.URL, "application/vnd.hbci", strings.NewReader(encodedMessage))
-	if err != nil {
-		return nil, err
-	}
-	defer httpResponse.Body.Close()
-	var marshaledResponse []byte
-	if httpResponse.StatusCode == http.StatusOK {
-		decodedReader := base64.NewDecoder(base64.StdEncoding, httpResponse.Body)
-		marshaledResponse, err = ioutil.ReadAll(decodedReader)
-		if err != nil {
-			return nil, err
-		}
-	} else {
-		marshaledResponse, err = ioutil.ReadAll(httpResponse.Body)
-		if err != nil {
-			return nil, err
-		}
-	}
-	return ReadResponse(marshaledResponse, request)
-}
-
-func NewRequest() *Request {
-	return &Request{}
-}
+type Middleware func(Transport) Transport
 
 type Request struct {
 	URL              string
